+++ ccn-lite/src/ccnl-core/include/ccnl-pkt.h
@@ -78,7 +78,7 @@ struct ccnl_pktdetail_ndntlv_s {
     int minsuffix, maxsuffix, mbf, scope;
     struct ccnl_buf_s *nonce;      /**< nonce */
     struct ccnl_buf_s *ppkl;       /**< publisher public key locator */
-    uint32_t interestlifetime;     /**< interest lifetime */
+    int64_t interestlifetime;     /**< interest lifetime */
     /* Data */
     uint32_t freshnessperiod;      /**< defines how long a node has to wait (after the arrival of this data before) marking it “non-fresh” */
 };
